+++ packages/jupyterlab-fonts/src/index.ts
@@ -93,6 +93,7 @@ export namespace DOM {
   export const sheet = 'jp-Fonts-Sheet';
   export const modGlobal = 'jp-fonts-mod-global';
   export const modNotebook = 'jp-fonts-mod-notebook';
+  export const notebookPanel = 'jp-NotebookPanel';
 }
 
 export type ICSSTextOptions = {