+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## v2.1.0 (unreleased)
+
+- Improve notebook-level `@import`, `@font-face`, etc.
+- adds `data-jpf-cell-id` and `data-jpf-cell-tags` to notebook cell elements
+
 ## v2.0.0
 
 - Support JupyterLab 3